requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Werkzeug==3.0.3
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.3
MarkupSafe==2.1.3
python-dotenv==1.0.0

# Image Matching Dependencies
# ResNet (Deep Learning) - PyTorch
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0

# SIFT (Feature Matching) - OpenCV
opencv-python>=4.8.0

# NumPy for numerical operations
numpy>=1.24.0

